[DEV-79] 시간표 추천 로직 리팩토링 (#356)
* refactor: Category(GraduationCategory) 단위 추천 → DetailCategory 단위 추천으로 고도화

* refactor: 전공 과목 수강가능학년 체크 로직 수정

* refactor: 시간표 추천의 채플 이수학점 체크를 int -> double

* feat: 인기과목을 학점 추천에 사용

* refactor: 시간표 과목 조회 무한스크롤 적용

* test: 무한 스크롤 적용으로 인한 테스트 코드 수정

* refactor: Lombok @Getter 적용

* refactor: 미사용 findByKeyword 메서드 제거

* test: findByYearAndSemesterAndNameContaining_filtersByNameKeyword 테스트 제거

* refactor: 메서드 분리로 복잡도 감소

* test: 시간표 추천 테스트 코드 작성

* refactor: TimetableController SonarCloud 이슈 수정

- 주석 코드 제거
- Stream.toList() 사용
- @ApiResponses 래퍼 제거
- 불필요한 import 정리

* refactor: addBasicCultureByMajor 메서드 분리

* refactor: SonarCloud 이슈 수정 (Math.clamp, Stream.toList, record 전환, 파라미터 감소)

* refactor: SonarCloud 이슈 수정 (Stream.toList(), 주석 코드 제거)

* refactor: SonarCloud 이슈 수정
- FindTimetableUseCase: searchCombinedWithPagination 파라미터 감소 (9개 → 3개)
  - SearchParams record 도입 (검색 조건 그룹화)
  - PaginationParams record 도입 (페이지네이션 파라미터 그룹화)

* refactor: SonarCloud 이슈 수정
- RecommendAfterTimetableService: 루프 내 break/continue 문 감소
  - processCategoryRequirements: continue 2개 → 1개로 감소
  - processBasicCultureLectures: continue/break 통합
  - addGlobalPopularLectures: shouldProcessPopularLecture 헬퍼 메서드 추출
- RecommendAfterTimetableServiceTest: AssertJ assertion 개선
  - containsEntry 사용 및 assertion 체이닝 적용

* test: 시간표 추찬 테스트 코드 추가

* test: 두 개의 assertion을 하나의 체인으로 통합

+++ src/main/java/com/plzgraduate/myongjigraduatebe/timetable/application/service/RequirementSnapshotQueryImpl.java
@@ -28,8 +28,8 @@ class RequirementSnapshotQueryImpl implements RequirementSnapshotQueryPort {
                     credit.getGraduationCategory(),
                     RequirementSnapshot.Item.builder()
                             .category(credit.getGraduationCategory())
-                            .totalCredit((int) credit.getTotalCredit())
-                            .takenCredit((int) credit.getTakenCredit())
+                            .totalCredit(credit.getTotalCredit())
+                            .takenCredit(credit.getTakenCredit())
                             .build()
             );
         }
